pyproject.toml
[project]
name = "trabalho2-persistencia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "faker>=38.2.0",
    "fastapi>=0.122.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
